Just pushing a new version of the star tracker.
