+++ starter-code/client/components/Popup.module.css
@@ -22,7 +22,6 @@
     color: gold;
   }
 
-
 #name{
     font-size: 20px;
     display: inline-block;
